requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
jinja2>=3.1.0
python-multipart>=0.0.9
pydantic-settings>=2.0.0
llama-index>=0.10.0
llama-index-llms-openai>=0.1.0
graspologic>=3.3.0
pandas>=2.0.0
networkx>=3.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pymupdf>=1.24.0
python-docx>=1.0.0
trafilatura>=1.8.0
beautifulsoup4>=4.12.0
nest-asyncio>=1.6.0
